Refactored test project.

+++ LLDB Test Application/LLDB iOS App/Classes/Controllers/PickersViewController.m
@@ -60,7 +60,7 @@
 #pragma mark - UIPickerViewDelegate
 - (NSString *)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component
 {
-    return [NSString stringWithFormat:@"%d", row];
+    return [NSString stringWithFormat:@"%ld", (long)row];
 }
